fixed some incorrect data types

+++ sear/irrseq00/irrseq00.hpp
@@ -323,9 +323,9 @@ typedef struct {
   // This data is in 3 bytes,
   // so I had to do this stupid solution to get it work
   unsigned int result_buffer_length : 24;
-  uint16_t bit_flags;
+  uint32_t bit_flags;
   char subsystem_prefix[8];
-  uint16_t rrsf_node_index;
+  uint32_t rrsf_node_index;
   uint8_t automatic_command_redirection;
   racf_rrsf_set_settings_t command_redirection_settings;
   uint8_t automatic_password_redirection;
@@ -334,7 +334,7 @@ typedef struct {
   racf_rrsf_set_settings_t password_synchronization_settings;
   uint8_t automatic_redirection_application_updates;
   racf_rrsf_set_settings_t application_updates_redirection_settings;
-  uint16_t number_of_rrsf_nodes;
+  uint32_t number_of_rrsf_nodes;
   char racf_subsystem_name[4];
   char racf_subsystem_userid[8];
   char reserved_space[52];
